Add space between view action buttons
View action buttons (create, export, filter, etc.) used to be glued to
one another in the production environment (because of uglifyjs
compression of html whitespace), but not in dev - that's why it went
unnoticed for so long.

While fixing the problem, I also rewrote the templates of the buttons
directives with the ES6 template literal syntax.